CRDL-264: Set up scala fmt, update libraries, refactoring and adds alias
