+++ .scalafmt.conf
@@ -1,0 +1,13 @@
+version = 3.9.6
+
+runner.dialect = scala3
+
+preset = defaultWithAlign
+
+maxColumn = 100
+
+indent {
+  callSite = 2
+  defnSite = 2
+  extendSite = 2
+}
