+++ README.md
@@ -1,7 +1,24 @@
 
 # crdl-ref-data-dps-stub
 
-This is a placeholder README.md for a new repository
+The Central Reference Data Data Platform Services(DPS) Stub responsibilities:
+* Simulate the responses of DPS
+
+## Prerequisites
+
+To ensure that you have all the prerequisites for running this service, follow the Developer setup instructions in the MDTP Handbook.
+
+This should ensure that you have the prerequisites for the service installed:
+
+* JDK 21
+* sbt 1.10.x or later
+* MongoDB 7.x or later
+* Service Manager 2.x
+
+### All tests and checks
+This is an sbt command alias specific to this project. It will run a scala format
+check, run unit tests, run integration tests and produce a coverage report:
+> `sbt runAllChecks`
 
 ### License
