+++ app/uk/gov/hmrc/crdlrefdatadpsstub/controllers/CodeListsController.scala
@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.crdlrefdatadpsstub.config
+package uk.gov.hmrc.crdlrefdatadpsstub.controllers
 
+import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
+import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import javax.inject.{Inject, Singleton}
-import play.api.Configuration
+import scala.concurrent.Future
 
-@Singleton
-class AppConfig @Inject() (config: Configuration) {
+@Singleton()
+class CodeListsController @Inject() (cc: ControllerComponents) extends BackendController(cc) {
 
-  val appName: String = config.get[String]("appName")
+  def hello(): Action[AnyContent] = Action.async { implicit request =>
+    Future.successful(Ok("Hello world"))
+  }
 }
