+++ conf/application.conf
@@ -1,3 +1,16 @@
+# Copyright 2025 HM Revenue & Customs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 include "backend.conf"
 
@@ -36,11 +49,6 @@ play.http.router = prod.Routes
 # The below controllers are the default exceptions to this rule.
 
 controllers {
-  uk.gov.hmrc.crdlrefdatadpsstub.controllers.MicroserviceHelloWorldController = {
-    needsAuth = false
-    needsLogging = false
-    needsAuditing = false
-  }
 }
 
 # Microservice specific config
@@ -48,10 +56,6 @@ controllers {
 
 microservice {
   services {
-    auth {
-      host = localhost
-      port = 8500
-    }
   }
 }
