+++ project/AppDependencies.scala
@@ -6,16 +6,14 @@ import sbt._
 object AppDependencies {
 
   private val bootstrapVersion = "9.11.0"
-  
 
-  val compile = Seq(
-    "uk.gov.hmrc"             %% "bootstrap-backend-play-30"  % bootstrapVersion
+  val compile: Seq[ModuleID] = Seq(
+    "uk.gov.hmrc" %% "bootstrap-backend-play-30" % bootstrapVersion
   )
 
-  val test = Seq(
-    "uk.gov.hmrc"             %% "bootstrap-test-play-30"     % bootstrapVersion            % Test,
-    
+  val test: Seq[ModuleID] = Seq(
+    "uk.gov.hmrc" %% "bootstrap-test-play-30" % bootstrapVersion % Test
   )
 
-  val it = Seq.empty
+  val it: Seq[ModuleID] = Seq.empty
 }
